fix: add tilted-chord dual-marker test and min-chord boundary test
Closes a review gap where no test exercised the actual rotation
direction of crop_and_rotate_roi() against real pixel content for a
genuinely tilted chord (dx != 0 and dy != 0) -- prior tests only
covered a vertical (no-rotation) chord marker, or checked arithmetic
hip_point/shoulder_point values independent of the rotation matrix.

Also simplifies the redundant int(round(...)) to round(...) for
out_w/out_h, and adds a test at the exact _MIN_CHORD_PX boundary.

+++ core/spine_contour.py
@@ -46,8 +46,8 @@ def crop_and_rotate_roi(frame: np.ndarray, hip_px: tuple[float, float],
     mid_x = (hip_x + shoulder_x) / 2.0
     mid_y = (hip_y + shoulder_y) / 2.0
 
-    out_w = int(round(chord_len * (1 + 2 * _ROI_MARGIN_FRAC)))
-    out_h = int(round(chord_len * (1 + 2 * _ROI_END_PAD_FRAC)))
+    out_w = round(chord_len * (1 + 2 * _ROI_MARGIN_FRAC))
+    out_h = round(chord_len * (1 + 2 * _ROI_END_PAD_FRAC))
 
     M = cv2.getRotationMatrix2D((mid_x, mid_y), angle_deg, 1.0)
     M[0, 2] += out_w / 2.0 - mid_x
